이메일 인증, 로그인 모듈화
